pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mcp2cli"
version = "0.1.0"
description = "Convert MCP server tools into hierarchical CLI commands and agent Skills"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "mengxiangyu" },
]
keywords = [
    "mcp",
    "cli",
    "agent",
    "skill",
    "model-context-protocol",
    "llm",
    "ai-tools",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries",
    "Topic :: Utilities",
]
dependencies = [
    "click>=8.1",
    "pyyaml>=6.0",
    "mcp>=1.0",
]

[project.urls]
Homepage = "https://github.com/mengxiangyu/mcp2cli"
Repository = "https://github.com/mengxiangyu/mcp2cli"
Issues = "https://github.com/mengxiangyu/mcp2cli/issues"

[project.scripts]
mcp2cli = "mcp2cli.main:cli"

[tool.hatch.build.targets.wheel]
packages = ["mcp2cli"]

[tool.hatch.build.targets.wheel.force-include]
"mcp2cli/presets" = "mcp2cli/presets"
